protein-data-bank/pyproject.toml
[project]
name = "protein_data_bank"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.12",
    "httpx>=0.28.1",
    "mcp>=1.1.2",
    "python-dotenv>=1.0.1",
]

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[tool.hatch.build.target.sdist.force-include]
"server.py" = "src/protein_data_bank/server.py"
